Patches to 4.10.04
git-svn-id: https://svn.ufz.de/svn/ogs/trunk/sources@4015 3895a583-e809-0410-9bb6-aa9463026377

+++ MSH/msh_core.cpp
@@ -17,7 +17,7 @@ Task:
 Programing:
 06/2005 WW Implementation
 **************************************************************************/
-CCore::CCore(const int id)
+CCore::CCore(const long id)
 {
   index = id;
   mark = true;
